Add files via upload
Firmware files for the SD File. ButtonSDTest05 tests the functionality of the SD card. SDTestWiFi tests the functionality of the SD card when the connection is lost. SDBatteryTest test the functionality of the SD card when main power is lost. arduino_secrets.h is where you will put down the name and password of your WiFi network.