Add files via upload
This is the latest version of the Arduino firmware that I've been working on. It includes the code to setup the Arduino and connect it to wifi, data collection capability, SD card capability, and power source detection. Need to test the firmware's ability to reconnect to WiFi, and add wireless commands.