Add files via upload
Main changes: Added the functions httpCommand and postDataToServer. 

httpCommand reads characters from an http client. Based on what the string is, the Arduino will carry out a specified function. Currently the only command is "rc", but more can be added as needed.

postDataToServer uses the POST method to post the most recent .csv file onto the server. 

Other changes include reducing the amount of global variables that are used in the firmware.